Added GR4J kernel

+++ hydrodiy/hydata/c_baseflow.h
@@ -3,5 +3,5 @@
 #include <stdio.h>
 #include <errno.h>
 
-int c_baseflow(int method, int nval, double* params, double* inputs, int* outputs);
+int c_baseflow(int method, int nval, double* params, double* inputs, double* outputs);
